requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.31.0
flask[async]==2.3.2
mysql-connector-python==8.1.0
pytz==2023.3.post1